Generator

Description

Generator

Documentation

sampleMaze :: Maze

Maze example

geraAleatorios :: Int -> Int -> [Int]

Given a seed returns a list of n integer randomly generated

nrAleatorio :: Int -> Int

Given a seed returns an integer randomly generated

subLista :: Int -> [a] -> [[a]]

Converssta list into a list of list of size n

convertePiece :: Int -> Piece

Converts an integer number into a Piece, 3 = Food Big 0 n < 70 <= Food Little 70 <= 99 <= Wall

printCorridor :: Corridor -> String

Converts a Corridor to a string

converteCorridor :: [Int] -> Corridor

Converts a list of integers into a Corridor

converteMaze :: [[Int]] -> Maze

Converts a list of lists of integers into a Maze

geraMaze :: Int -> Int -> Int -> Maze

Generates a Maze using a random seed