Tarefa1

Description

Introdução

Esta tarefa teve como objetivo a implementação de um mecanismo de geração de labirintos pseudo-aleatórios válidos.

Objetivos

Sem termos qualquer experiência de programação, de inicio foi extremamente complicado entrar no ritmo de trabalho, mas à medida que as aulas iam passado era notório que cada vez conseguiamos avançar mais na tarefa.

O planeamneto desta tarefa consistiu na separação do trabalho em partes cruciais. Por exemplo, primeiro foram tratados os limites do labirinto, depois seguiu-se a implementação dos tunéis e por último, que foi também onde tivemos mais dificuldade, a implementação da casa dos fantasmas.

As estratégias para ambas as partes foram similares. Sendo um labirinto nada mais que uma lista de listas de peças, o nosso foco sempre passou por trabalhar sempre que possivel num sistema de posições dentro da lista.

Discussão e conclusão

Apesar da dificuldade inicial, muito devida à nossa inexperiência na área, achamos ter terminado com sucesso a realização da Tarefa 1. Através de ínumeros testes com diferentes medidas de comprimento e largura conseguimos sempre visualizar um labirinto válido, com particular atenção para a casa dos fantasmas sempre no sitio certo, tal como os túneis.

Para concluir, achamos a tarefa um agradável desafio para introdução do projeto. E o bom resultado alcançado no final da tarefa acabou por nos transmitir confiança para as tarefas seguintes. Foi também importante estarmos em contacto com certas funções que apareceram no módulo Generator para também começar a tentar perceber o funcionamento de algo que consideramos avançado tendo em conta o estado da aprendizagem em que nos encontrávamos.

Funções gerais da Tarefa 1

donaldTrump :: Int -> Corridor

Cria um corredor com determinado numero de peças Wall

fronteiraMexicana :: Maze -> Corridor -> Maze

Dado um Maze devolve-o com limitado inferiormente por peças Wall

geraMazeAmericano :: Int -> Int -> Int -> Maze

Gera um maze limitado inferior e superiormente por peças Wall

bobConstrutor :: Int -> Maze

Casa dos fantasmas construida manualmente, com peças Empty à volta

exclamacao :: [[Piece]] -> Int -> [Piece]

Definição recursiva da função (!!) já definida no Prelude

ifEvenA :: Maze -> Bool

Testa a paridade da medida da altura de um Maze

ifEvenC :: Maze -> Bool

Testa a paridade da medida da largura de um Maze

getmiddleCorridor :: Maze -> [Corridor]

Dado um Maze devolve o corredor central quando este tem altura impar ou os dois centrais quando a altura é par

replace :: Maze -> Maze

Dado um Maze devolve-o limitado à esquerda por peças Wall

replacelast :: Maze -> Maze

Dado um Maze devolve-o limitado à direita por peças Wall

mazecomparedes :: Int -> Int -> Int -> Maze

Gera um Maze totalmente rodeado por peças Wall

Implementação dos túneis

tunellast :: Maze -> [Corridor]

Devolve o corredorcorredores centrais de um Maze com umdois tuneis no final

tunelbegin :: Maze -> [Corridor]

Devolve o corredorcorredores centrais de um Maze com umdois tuneis no inicio

mazecomtunelpar :: Maze -> Maze

Devolve um Maze (par) com dois tuneis no fim

mazecomtunelcompletopar :: Maze -> Maze

Devolve um Maze (par) com dois tuneis no inicio

mazequasecompletopar :: Int -> Int -> Int -> Maze

Gera um Maze (par) rodeado por Wall com tuneis nos dois lados

mazecomtunelimpar :: Maze -> Maze

Devolve um Maze (impar) com dois tuneis no fim

mazecomtunelcompletoimpar :: Maze -> Maze

Devolve um Maze (impar) com dois tuneis no inicio

mazequasecompletoimpar :: Int -> Int -> Int -> Maze

Gera um Maze (impar) rodeado por Wall com tuneis nos dois lados

typemaze :: Int -> Int -> Int -> Maze

Gera um Maze (par ou impar) rodeado por Wall com tuneis nos dois lados

Implementação da casa dos fantasmas

get5middlecorridor :: Maze -> [Corridor]

Devolve os 5 corredores centrais de um labirinto (par ou impar)

replaceinMaze :: Maze -> Maze -> Maze

Substitui os 5 corredores centrais de um Maze

middle3Corridors :: Maze -> Maze

Implementa a casa dos fantasmas no labirinto

mazecompleto :: Maze -> Maze

Devolve um Maze (par ou impar) com a casa dos fantasmas no meio

Função final da Tarefa 1

generateMaze :: Int -> Int -> Int -> Maze

Gera um Maze pseudo-aleatório com as dimensões desejadas

Funções de teste

alturacerta :: Int -> Int -> Int -> Bool

Teste altura

larguracorreta :: Int -> Int -> Int -> Bool

Teste largura

onlyWalltop :: Int -> Int -> Int -> Bool

Teste (Maze limitado superiormente por Wall)

onlyWallbottom :: Int -> Int -> Int -> Bool

Teste (Maze limitado inferiormente po Wall)

tunnelpar :: Int -> Int -> Int -> Bool

Teste (Maze de altura par && tuneis no meio)

tunnelimpar :: Int -> Int -> Int -> Bool

Teste (Maze de altura impar && tunel no meio)