Tarefa2

Description

Introdução

Esta Tarefa teve como objetivo a determinação do efeito de uma jogada no estado do jogo. Esta tarefa acabou por se revelar a mais trabalhosa, no sentido que foi preciso atentar não só no encunciado, mas também em vários pormenores do jogo real na tentativa de conhecer todas as jogadas e acontecimentos possíveis. Foi também necessária uma remodelação posteriror à entrega da primeira fase, de forma a poder continuar a trabalhar sobre as tarefas da seunda fase.

Objetivos

Tal como na Tarefa 1, optamos por seguir uma estratégia de separação do trabalho por partes. A primeira dificuldade passou por compreender o novo data type, e todos os novos tipos que apareceram. Para além disso foi necessário criar um novo data Type nosso de forma a facilitar o controlo dos diferentes tipos de acontecimentos que uma jogada pode ter.

Dado que o resultado final é um State alterado conforme o tipo de jogada, foi prontamente criada uma função que deteta essa mesma ação dentro dos acontecimentos definidos no PlayAction, e foi a partir daí que começou a parte mais trabalhosa da tarefa. Foi preciso definir os vários conportamentos nos Players desde a alteração dos estados, do número de vidas e pontos no Pacman, e para além das alterações próprias da tarefa, optamos por implementar também algumas das alteraçõe gráficas que ocorrem em cada ação, tais como a aparição de uma peça vazia após o Pacman comer por exemplo uma comida pequena.

Sem dúvida que o que nos causou mais dificuldade foi a interação do Pacman com os fantasmas. Acabamos por demorar algum tempo a conseguir completar a nossa função de deteção de forma a detetar tanto estados dos jogadores, como jogadores em si.

Discussão e conclusão

Apesar de ser a mais trabalhosa tarefa da primeira fase, acabou por ser também a que nos deu mais prazer em fazer. Foi uma tarefa feita de passo em passo, e que cada pequena vitória nos dava cada vez mais motivação.

Mais uma vez conseguimos alcançar o que pretendíamos, e com a ajuda dos testes quer feitos por nós ou os disponibilizados no codeboard, percebemos que realmente não acontecia nada que não fosse suposto. Testamos praticamente todas as situações possiveis de acontecimentos, desde a simples movimentação para um dos lados, até ao diferentes tipos de choques com fantasmas ou a passagem pelos túneis, e tudo funcionou como era esperado.

Para concluir, temos a completa noção de que a forma como estrutuámos o código não foi das mais efecientes, mas a verdade é que ainda não tinhamos grande conhecimento e dominio sobre a linguagem, e achamos mais importante salientar que todos os pontos do enunciado foram cobertos e tal como na tarefa 1 acabamos a tarefa com a sensação de trabalho realizado.

Tipos criados de auxílio à Tarefa 2

data PlayAction

Tipos de acontecimentos numa Play

Constructors

Eat EatType

Collide

Die

DieG

TeleportL

TeleportR

MoverVazio

DirecaoMuda

data EatType

Tipos de comida

Constructors

Pequena

Grande

Fantasma

Jogador

Inferir a ação

getPiece :: [Player] -> Coords -> Maze -> Piece

Devolve a peça de determinadas coordenadas de um Maze com jogadores

moveCoords :: State -> Play -> Coords

Infere a posição final do Player terminada a Play

isPiecePacmanMega :: Piece -> Bool

Infere se uma peça é um Pacman em estado Mega

isPiecePacman :: Piece -> Bool

Infere se uma peça é um Pacman

getMovePiece :: Play -> State -> Piece

Devolve a peça que se encontra na posição para a qual o Player se desloca

acaoPiece :: Play -> State -> PlayAction

Dada uma Play e um State infere o tipo de ação

infereAcao :: Player -> Piece -> Maze -> PlayAction

Infere o tipo de ação dada a peça para que o Player se move

Atualização dos jogadores

setPlayersPoints :: Player -> PlayAction -> Player

ALtera os pontos de um Player conforme a PlayAction

setPlayersLives :: Player -> PlayAction -> Player

Altera o número de vidas de um Player conforme a PlayAction

setPlayerState :: Player -> PlayAction -> Player

Altera o estado de um Player conforme a PlayAction

getPlayerLives :: Player -> Int

Devolve o numero de vidas de um Player

choosePlayer :: Int -> [Player] -> Player

Determina o jogador a aplicar a jogada a partir do ID

choosePlayerWithCoords :: Coords -> [Player] -> Player

Devolve um jogador de uma lista com determinadas coordenadas

removeElem :: Play -> [Player] -> [Player]

Remove o Player da lista (Tarefa2)

changeOrientation :: Orientation -> Player -> Player

Muda a orientação de um Player

ghoDeadOrientation :: Player -> Player

Inverte a orientação de um Ghost

updatePlay :: Play -> Maze -> PlayAction -> [Player] -> [Player]

Dada uma PlayAction devolve a lista dos jogadores alterada conforme o tipo da mesma

getPlayerGhostPlay :: Maze -> Play -> [Player] -> Player

Devolve o Pacman envolvido na jogada do fantasma

updatePlayFantasmas :: Maze -> Play -> [Player] -> Player

Atualiza o estado do fantasma que é comido

Atualização do labirinto

updateMaze :: Player -> PlayAction -> Maze -> Maze

Dada uma PlayAction altera o Maze conforme o tipo da mesma

Função final da Tarefa 2

play :: Play -> State -> State

Efetua uma jogada para um determinado jogador

Funções de teste

sameMode :: [Player] -> Bool

Testa se os fantasmas de uma lista têm o mesmo estado

moveright :: Play -> State -> Bool

Teste (movimentação simples para a direita)

moverleft :: Play -> State -> Bool

Teste (movimentação simples para a esquerda)

moverdown :: Play -> State -> Bool

Teste (movimentação simples para baixo)

moverup :: Play -> State -> Bool

Teste (movimentação simples para cima)

colisao :: Play -> State -> Bool

Teste (colisão com parede)

tuneldireita :: Play -> State -> Bool

Teste (passar túnel da direita)

tunelesquerda :: Play -> State -> Bool

Teste (passar tunel da esquerda)

tunelvoltar :: Play -> State -> Bool

Teste (passar tunel da direita e voltar)

comerpequena :: Play -> State -> Bool

Teste (mover para a direita e comer comida pequena)

comergrande :: Play -> State -> Bool

Teste (mover para a esquerda e comer comida grande && fantasmas em modo Dead)

baterfantasma :: Play -> State -> Bool

Teste (colisão com fantasma Alive e perder uma vida)

morrer :: Play -> State -> Bool

Teste (ter 0 vidas e colisao com fantasma e passar para modo Dying)

comerfantasma :: Play -> State -> Bool

Teste (mover para baixo e comer fantasma && pontuação certa && fantasma renasce na casa em modo Alive)

comerfantasmaEcomidapequena :: Play -> State -> Bool

Teste (mover para baixo e comer fantasma em casa com peça Food Little && pontuação certa && fantasma renasce na casa em modo Alive)

comerfantasmaEcomidagrande :: Play -> State -> Bool

Teste (mover para baixo e comer fantasma em casa com peça Food Big && pontuação certa && fantasma renasce na casa em modo Alive)