Tarefa5

Description

Introdução

Esta tarefa teve como objetivo a criação de jogadas para os fantasmas. Foi uma tarefa desafiante, pelo que o mais importante a ter em conta foi quais são as melhores reações dos fantasmas ao pacman e criar uma estratégia para eles executarem.

Objetivos

Compreender o objetivo da tarefa não foi muito complicado. Todos os data types que foram necessários para a completar já tinham sido anteriormente definidos em tarefas anteriores, aos quais já estávamos portanto habituados a utilizar (principalmente desde a tarefa 2).

Sabendo que todas as jogadas dos fantasmas em modo Alive são reações ao pacman, tivemos que criar uma grande quantidade de funções que comparam coordenadas de jogadores e a localização deles no labirinto. O nosso objetivo foi fazer os fanstamas perseguir constantemente o pacman pelo melhor caminho possível. O mais desafiante nesse aspecto, foi fazer os fantasmas reagir ao tunel no caso em que passar pelo mesmo seja a melhor jogada para apanhar o pacman. Para tal, foi necessário ter em conta uma grande quantidade de situações diferentes em conta, o que tornou o código desta tarefa mais complexo.

No modo Dead, os fantasmas não são muito reactivos, visto que apenas seguem um caminho em linha reta até se encontrarem com uma parede, o que os faz virar para a sua direita e continuar o caminho em linha reta. Visto que é uma movimentação bastante simples, foi muito provavelmente a parte mais fácil de codificar nesta tarefa.

Para inverter a direção dos fantasmas quando eles entram em modo Dead, fizemos umas alterações na tarefa 2. Foi adicionada uma função auxiliar que inverte a direção de um fantasma, tendo usado a mesma no caso em que o pacman come uma Comida Grande.

Discussão e conclusão

Embora tenha tido alguns desafios, esta tarefa deixou-nos bastante entusiasmados, principalmente tendo em conta que foi ela que tornou o nosso jogo no seu estado final, onde todos os elementos estão funcionais.

Apesar disso, alguns aspectos desta tarefa não ficaram exatamente como nós gostaríamos que tivessem ficado, tendo sido o principal motivo o curto tempo que conseguimos dedicar à mesma. Um exemplo é a codificação desta fase do trabalho, que poderia provavelmente estar mais compacta e eficaz.

Para concluir, embora não esteja perfeita, esta tarefa cumpriu o seu objetivo de criar jogadas para os fantasmas, o que já foi um motivo de contentamento.

Funções gerais da Tarefa 5

pCoords :: [Player] -> [Coords]

Lista das coordenadas de lista de jogadores

difCoordsI :: Player -> Player -> Coords

Distancia entre dois players em coordenadas

difCoords :: Player -> Player -> Coords

Distancia entre um fantasma e as coordenadas para onde ele vai para apanhar o pacman

difCoordsC :: Coords -> Coords -> Coords

Distancia em coordenadas entre uma coordenade e um player

difCoordsHT :: Coords -> Player -> Maybe Int

Distancia horizontal entre uma coordenada e um player (para usar com o tunel)

Funções de tracking (Onde se encontra um elemento em relação a um jogador no labirinto)

whereTunH :: Player -> Player -> Maze -> Orientation

Mostra se o tunel estã à direita ou à esquerda de um jogador

whereTunV :: Player -> Player -> Maze -> Orientation

Mostra se o tunel está acima ou abaixo de um jogador

pertoDoTunel :: Player -> Player -> Maze -> Coords

Dá as coordenadas do tunel no qual o fantasma tem que entrar no caso de ser a melhor jogada para apanhar o pacman

wherePacH :: Player -> Player -> Orientation

Mostra se um player está à direita ou à esquerda de outro jogador

wherePacV :: Player -> Player -> Orientation

Mostra se um player está acima ou abaixo de outro jogador

Funções de decisão de jogadas

basicGPlayA :: Player -> Player -> Maze -> Play

Decide a jogada para um fantasma apanhar o pacman.

collideGPlayA :: Player -> Player -> Play

Decide a jogada para um fantasma apanhar o pacman no caso de ele se encontrar com uma parede.

decideGPlayA :: Player -> Player -> State -> Play

Decide a jogada final para o fantasma apanhar o pacman

chaseMode :: State -> Int -> Play

Jogada do fantasma para apanhar o pacman

basicGPlayD :: Player -> Play

Jogada basica do fantasma quando ele está em modo Dead

collideGPlayD :: Player -> Play

Jogada do fantasma em modo Dead quando ele se encontra com uma parede

decideGPlayD :: Player -> State -> Play

Decide a jogada final para o fantasma em modo Dead

scatterMode :: State -> Int -> Play

Jogada do fantasma em modo Dead

playToList :: [Player] -> State -> [Play]

Cria uma lista de jogadas, uma para cada fantasma

Função final da Tarefa 5

ghostPlay :: State -> [Play]

Devolve um conjunto de jogadas, uma para cada fantasma